requirements.txt
numpy==1.26.0
pandas==2.1.1
geopandas==0.14.0
psycopg2-binary==2.9.7
twilio==8.9.1
python-dotenv==1.0.0
